pyproject.toml
[tool.poetry]
name = "price_checker"
version = "0.1.0"
description = ""
authors = ["pavelborisofff <pavel.borisofff@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.6.0"
pydantic = "^1.10.4"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

